main-service/requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
grpcio==1.35.0
grpcio-tools==1.35.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.14.0
pylint==2.6.0
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
